requirements.txt
pillow==10.2.0
praw==7.7.1
python-dotenv==1.0.1
requests==2.31.0
pyperclip==1.8.2

